feat(src): __main__ is used solely to start bot
