+++ src/__main__.py
@@ -1,0 +1,15 @@
+import asyncio
+import logging
+import sys
+
+from src.loader import bot
+from src.loader import dp
+
+
+async def main() -> None:
+    await dp.start_polling(bot)
+
+
+if __name__ == "__main__":
+    logging.basicConfig(level=logging.INFO, stream=sys.stdout)
+    asyncio.run(main())
